Implement Cluster.distance functions
Note that the algorithmic complexity of these functions is O(nm), where
n is the length of self.points, and m is the length of other.points.

Accordingly, don't expect great results with large numbers of points?

Or, conversely, be prepared for a great amount of computation to occur
before getting that answer you're looking for?